+++ packages/s2-core/src/utils/export/copy/common.ts
@@ -77,7 +77,9 @@ export const assembleMatrix = ({
   const dataWidth = dataMatrix[0]?.length ?? 0;
   const dataHeight = dataMatrix.length ?? 0;
   const matrixWidth = rowWidth + dataWidth;
-  const matrixHeight = colHeight + dataHeight;
+  // 存在行头节点时，则必然存在行头标题的角头
+  const cornerHeight = colHeight || (rowWidth > 0 ? 1 : 0);
+  const matrixHeight = cornerHeight + dataHeight;
 
   let matrix: SimpleData[][] = Array.from(
     Array(matrixHeight),
@@ -86,7 +88,7 @@ export const assembleMatrix = ({
 
   matrix = map(matrix, (heightArr, y) =>
     map(heightArr, (_, x) => {
-      if (x >= 0 && x < rowWidth && y >= 0 && y < colHeight) {
+      if (x >= 0 && x < rowWidth && y >= 0 && y < cornerHeight) {
         return cornerMatrix?.[y]?.[x] ?? '';
       }
 
@@ -94,17 +96,17 @@ export const assembleMatrix = ({
         return colMatrix[y][x - rowWidth];
       }
 
-      if (x >= 0 && x < rowWidth && y >= colHeight && y < matrixHeight) {
-        return rowMatrix?.[y - colHeight][x];
+      if (x >= 0 && x < rowWidth && y >= cornerHeight && y < matrixHeight) {
+        return rowMatrix?.[y - cornerHeight][x];
       }
 
       if (
         x >= rowWidth &&
         x <= matrixWidth &&
-        y >= colHeight &&
+        y >= cornerHeight &&
         y < matrixHeight
       ) {
-        return dataMatrix[y - colHeight][x - rowWidth];
+        return dataMatrix[y - cornerHeight][x - rowWidth];
       }
 
       return undefined;
@@ -217,7 +219,11 @@ export const getNodeFormatData = (leafNode: Node) => {
     const formatter = node.spreadsheet?.dataSet?.getFieldFormatter?.(
       node.field,
     );
-    const value = formatter?.(node.value) as string;
+
+    const value =
+      node.isTotalRoot || !formatter
+        ? node.value
+        : (formatter(node.value) as string);
 
     line.unshift(value);
 
@@ -240,3 +246,5 @@ export const getHeaderNodeFromMeta = (
 
   return [facet.getRowNodeByIndex(rowIndex), facet.getColNodeByIndex(colIndex)];
 };
+
+export const ricOptions = { timeout: 3000 };